+++ android/app/src/main/java/com/orang1ver/yuanqiledger/MainActivity.java
@@ -1,0 +1,5 @@
+package com.orang1ver.yuanqiledger;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}